server/weather_api/requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
Flask==3.1.0
h3==4.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
pycparser==2.22
requests==2.32.3
sqlparse==0.5.3
timezonefinder==6.5.9
urllib3==2.4.0
Werkzeug==3.1.3
